Introduced arccps in the middle between arcmac & arcc.
Now some tests do not pass (have to do with apply, which needs fixing).
Added some tests for this.

The biggest problem (which made many tests fail) was 'quote'.
Removed quote from 'def-arc-walker', so now each walker deals with it
directly (arccps has to funcall 'cc', that was the problem).